Move console logic to StartSearchController
Bugfix: switches console output to the selected search.
