+++ client/src/main/resources/views/workspace/ConsoleView.fxml
@@ -29,8 +29,7 @@
 <AnchorPane fx:id="consoleArea" prefHeight="400.0" prefWidth="Infinity" styleClass="console-pane"
             minWidth="0" minHeight="0"
             xmlns="http://javafx.com/javafx/8.0.121"
-            xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="org.iconic.workspace.console.ConsoleController">
+            xmlns:fx="http://javafx.com/fxml/1">
     <ListView fx:id="consoleContent" styleClass="console-content"
               AnchorPane.bottomAnchor="0" AnchorPane.leftAnchor="0"
               AnchorPane.rightAnchor="0" AnchorPane.topAnchor="0" />
